BL-6085 refactor epub publish into its own class
