+++ src/BloomBrowserUI/publish/epub/epubVisibility.less
@@ -1,0 +1,20 @@
+// Items hidden only in epubs are defined here.
+// One might think they should be defined in baseEPUB.less.
+// But baseEPUB.less is actually shipped with epubs whereas basePage.less is used at the
+// time the epub is created to determine what elements should be shown and hidden.
+// Hidden elements will actually be removed from the dom
+// completely since some eReaders don't handle display:none properly.
+// The epub-visibility class and this stylesheet are added in EpubMaker.RemoveUnwantedContent()
+// and are *only* used to strip out elements we don't want included in our epub dom.
+.epub-visibility {
+    .credits {
+        .licenseAndCopyrightBlock {
+            .licenseUrl {
+                display: none; // included in the licenseDescription field, no need to repeat
+            }
+        }
+        .ISBNContainer {
+            display: none; // would probably need a separate number for epub anyway
+        }
+    }
+}
